Refactor the device factory
* Remove unnecessary types from the API package
* Avoid unncessary copying of bytes into Strings
