+++ prototype/southbound/device-factory/device-factory-core/src/main/java/org/eclipse/sensinact/gateway/southbound/device/factory/EncodingUtils.java
@@ -12,7 +12,8 @@
 **********************************************************************/
 package org.eclipse.sensinact.gateway.southbound.device.factory;
 
-import java.util.Arrays;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
 
 /**
  *
@@ -20,24 +21,31 @@ import java.util.Arrays;
 public class EncodingUtils {
 
     /**
-     * Removes the UTF-8 Byte Order Mark from the given input. Make sure the
-     * expected encoding is UTF-8 before calling this method
+     * Creates an input stream from the supplied byte array, removing the UTF-8 Byte
+     * Order Mark from the given input if present. Make sure the expected encoding
+     * is UTF-8 before calling this method
      *
      * @param rawInput Bytes of an UTF-8 content
      * @return The content without the BOM prefix
      */
-    public static byte[] removeBOM(final byte[] rawInput) {
-        if (rawInput != null && rawInput.length > 3) {
+    public static InputStream removeBOM(final byte[] rawInput) {
+        byte[] toCheck = rawInput == null ? new byte[0] : rawInput;
+
+        int offset = 0;
+        int size = toCheck.length;
+
+        if (rawInput.length > 3) {
             final int char0 = rawInput[0] & 0xFF;
             final int char1 = rawInput[1] & 0xFF;
             final int char2 = rawInput[2] & 0xFF;
 
             if (char0 == 0xEF && char1 == 0xBB && char2 == 0xBF) {
                 // Got an UTF-8 with BOM, remove the BOM marker
-                return Arrays.copyOfRange(rawInput, 3, rawInput.length);
+                offset = 3;
+                size -= 3;
             }
         }
 
-        return rawInput;
+        return new ByteArrayInputStream(toCheck, offset, size);
     }
 }
