+++ prototype/southbound/device-factory/device-factory-core/src/main/java/org/eclipse/sensinact/gateway/southbound/device/factory/impl/FactoryParserHandler.java
@@ -46,14 +46,9 @@ import org.eclipse.sensinact.gateway.southbound.device.factory.IPlaceHolderKeys;
 import org.eclipse.sensinact.gateway.southbound.device.factory.IResourceMapping;
 import org.eclipse.sensinact.gateway.southbound.device.factory.InvalidResourcePathException;
 import org.eclipse.sensinact.gateway.southbound.device.factory.MissingParserException;
-import org.eclipse.sensinact.gateway.southbound.device.factory.NamingUtils;
 import org.eclipse.sensinact.gateway.southbound.device.factory.ParserException;
 import org.eclipse.sensinact.gateway.southbound.device.factory.RecordPath;
-import org.eclipse.sensinact.gateway.southbound.device.factory.ResourceLiteralMapping;
-import org.eclipse.sensinact.gateway.southbound.device.factory.ResourceMappingHandler;
-import org.eclipse.sensinact.gateway.southbound.device.factory.ResourceRecordMapping;
 import org.eclipse.sensinact.gateway.southbound.device.factory.VariableNotFoundException;
-import org.eclipse.sensinact.gateway.southbound.device.factory.VariableSolver;
 import org.eclipse.sensinact.gateway.southbound.device.factory.dto.DeviceMappingConfigurationDTO;
 import org.eclipse.sensinact.gateway.southbound.device.factory.dto.DeviceMappingOptionsDTO;
 import org.eclipse.sensinact.prototype.PrototypePush;
